+++ scripts/simulate_sparsity_sweep_mpi.py
@@ -1,4 +1,5 @@
 import numpy as np
+from numpy.random import default_rng
 import pathlib
 from mpi4py import MPI
 
@@ -12,9 +13,10 @@ from causal_optoconnectics.generator import (
     generate_regular_stim_times,
     generate_oscillatory_drive,
     dales_law_transform,
+    sparsify
 )
 
-def construct(params, sparsity=0):
+def construct(params, rng=rng):
     # set stim
     binned_stim_times = generate_poisson_stim_times(
         params['stim_period'],
@@ -29,15 +31,7 @@ def construct(params, sparsity=0):
     )
     stimulus = np.concatenate((binned_stim_times, binned_drive), 0)
     W_0 = construct_connectivity_matrix(params)
-    indices = np.unravel_index(
-        np.random.choice(
-            np.arange(np.prod(W_0.shape)),
-            size=int(sparsity*np.prod(W_0.shape)),
-            replace=False
-        ),
-        W_0.shape
-    )
-    W_0[indices] = 0
+    W_0 = sparsify(W_0, params['sparsity'], rng=rng)
     W_0 = dales_law_transform(W_0)
     W, W_0, excit_idx, inhib_idx = construct_connectivity_filters(W_0, params)
     W = construct_additional_filters(
@@ -46,14 +40,14 @@ def construct(params, sparsity=0):
     W = construct_additional_filters(
         W, range(len(W_0)), params['drive_scale'], params['drive_strength'])
 
-    return W, W_0, stimulus
+    return W, W_0, stimulus, excit_idx, inhib_idx
 
 if __name__ == '__main__':
     data_path = pathlib.Path('datasets/sweep_3')
     data_path.mkdir(parents=True)
     comm = MPI.COMM_WORLD
     rank = comm.Get_rank()
-    np.random.seed()
+
     params = {
         'const': 5,
         'n_neurons': 50,
@@ -77,24 +71,28 @@ if __name__ == '__main__':
             'mu': 0,
             'sigma': 5
         },
-        'n_time_step': int(1e6)
+        'n_time_step': int(1e6),
+        'seed': 12345 + rank
     }
+    rng = default_rng(params['seed'])
+
     sparsities = [0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9]
-    if rank == 0:
-        connectivity = {}
-    else:
-        connectivity = None
+    connectivity = {}
+
     for sparsity in sparsities:
         path =  f'sparsity_{sparsity:.1f}'.replace('.','')
         (data_path / path).mkdir(exist_ok=True)
+        fname = data_path / path/ f'rank_{rank}.npz'
+        if fname.exists():
+            continue
+
         if rank == 0:
-            W, W_0, stimulus = construct(params, sparsity=sparsity)
-            connectivity[path] = (W, W_0, stimulus)
+            connectivity[path] = construct(params, sparsity=sparsity)
         connectivity = comm.bcast(connectivity, root=0)
-        W, W_0, stimulus = connectivity[path]
-        res = simulate(W=W, W_0=W_0, inputs=stimulus, params=params)
+        W, W_0, stimulus, excit_idx, inhib_idx = connectivity[path]
+        res = simulate(W=W, W_0=W_0, inputs=stimulus, params=params, rng=rng)
         np.savez(
-            data_path / path/ f'rank_{rank}',
+            fname,
             data=res,
             W=W,
             W_0=W_0,
