+++ scripts/simulate_sweep_mpi.py
@@ -1,4 +1,5 @@
 import numpy as np
+from numpy.random import default_rng
 import pathlib
 from mpi4py import MPI
 
@@ -14,12 +15,13 @@ from causal_optoconnectics.generator import (
     dales_law_transform,
 )
 
-def construct(params):
+def construct(params, rng):
     stimulus = generate_poisson_stim_times(
         params['stim_period'],
         params['stim_isi_min'],
         params['stim_isi_max'],
-        params['n_time_step']
+        params['n_time_step'],
+        rng=rng
     )
     W_0 = construct_connectivity_matrix(params)
     W_0 = dales_law_transform(W_0)
@@ -28,14 +30,14 @@ def construct(params):
         W, excit_idx[:params['n_stim']], params['stim_scale'],
         params['stim_strength'])
 
-    return W, W_0, stimulus
+    return W, W_0, stimulus, excit_idx, inhib_idx
 
 if __name__ == '__main__':
     data_path = pathlib.Path('datasets/sweep_1')
     data_path.mkdir(parents=True, exist_ok=True)
     comm = MPI.COMM_WORLD
     rank = comm.Get_rank()
-    np.random.seed()
+
     params = {
         'const': 5,
         'n_neurons': None,
@@ -56,15 +58,13 @@ if __name__ == '__main__':
             'mu': 0,
             'sigma': None
         },
-        'n_time_step': int(1e6)
+        'n_time_step': int(1e6),
+        'seed': 12345 + rank
     }
     n_neuronss = [10, 20, 30, 40, 50]
     stim_strengths = [1, 2, 3, 4, 5, 6, 7, 8]
     sigmas = [0.5, 1, 2, 3, 4, 5, 6, 7]
-    if rank == 0:
-        connectivity = {}
-    else:
-        connectivity = None
+    connectivity = {}
     for n_neurons in n_neuronss:
         for stim_strength in stim_strengths:
             for sigma in sigmas:
@@ -74,14 +74,18 @@ if __name__ == '__main__':
                 params['n_neurons'] = n_neurons
                 path =  f'n{n_neurons}_ss{stim_strength}_s{sigma}'.replace('.','')
                 (data_path / path).mkdir(exist_ok=True)
+                fname = data_path / path/ f'rank_{rank}.npz'
+                if fname.exists():
+                    continue
+
                 if rank == 0:
-                    W, W_0, stimulus = construct(params)
-                    connectivity[path] = (W, W_0, stimulus)
+                    connectivity[path] = construct(params)
                 connectivity = comm.bcast(connectivity, root=0)
-                W, W_0, stimulus = connectivity[path]
-                res = simulate(W=W, W_0=W_0, inputs=stimulus, params=params)
+
+                W, W_0, stimulus, excit_idx, inhib_idx = connectivity[path]
+                res = simulate(W=W, W_0=W_0, inputs=stimulus, params=params, rng=rng)
                 np.savez(
-                    data_path / path/ f'rank_{rank}',
+                    fname,
                     data=res,
                     W=W,
                     W_0=W_0,
